third_party/WebKit/LayoutTests/fast/css/content-distributed-nodes-expected.txt

Checking if styles in the nested shadow roots apply properly to distributed elements. (v0 bug compatible)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(span).color is "rgb(255, 0, 0)"
PASS getComputedStyle(span).backgroundColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
red?
